Did some code cleaning and refactoring.

+++ packages/laravel-issue-tracker/metadata/src/Controllers/MetadataController.php
@@ -1,40 +1,42 @@
 <?php
+
 namespace LaravelIssueTracker\Metadata\Controllers;
 
-use Illuminate\Support\Facades\Request;
-use LaravelIssueTracker\Core\Acme\Validators\ValidationException;
 use Illuminate\Support\Facades\Input;
+use Illuminate\Support\Facades\Request;
+use LaravelIssueTracker\Metadata\Models\Metadata;
 use LaravelIssueTracker\Core\Controller\ApiController;
+use LaravelIssueTracker\Core\Acme\Validators\ValidationException;
 use LaravelIssueTracker\Metadata\Acme\Services\MetadataCreatorService;
 use LaravelIssueTracker\Metadata\Acme\Transformers\MetadataTransformer;
-use LaravelIssueTracker\Metadata\Models\Metadata;
 
 class MetadataController extends ApiController {
 
     /**
-     * @var
-     */
-    protected $metadataTransformer;
-
-    /**
-     * @var
+     * Property for metadata creator service.
      */
     protected $metadataCreator;
 
     /**
+     * Property for the metadata transformer.
+     */
+    protected $metadataTransformer;
+
+    /**
      * MetadataController constructor.
+     *
      * @param $metadataTransformer
      * @param $metadataCreator
      */
     public function __construct(MetadataTransformer $metadataTransformer, MetadataCreatorService $metadataCreator)
     {
-        $this->metadataTransformer = $metadataTransformer;
         $this->metadataCreator = $metadataCreator;
+        $this->metadataTransformer = $metadataTransformer;
     }
 
 
     /**
-     * Display a listing of the resource.
+     * Display the list of the resource.
      *
      * @return Response
      */
@@ -60,8 +62,9 @@ class MetadataController extends ApiController {
     public function store()
     {
         try {
-            //$this->authorize('store');
+            $this->authorize('store');
             $this->metadataCreator->make(Input::all());
+
             return $this->respondCreated('Metadata successfully created!');
         }
         catch (ValidationException $e) {
@@ -77,15 +80,13 @@ class MetadataController extends ApiController {
      */
     public function show($id)
     {
-        $metadata = Metadata::find($id);
-
-        if( ! $metadata )
+        if( ! Metadata::find($id)->exists() )
         {
             return $this->respondNotFound('Metadata does not exist');
         }
 
         return $this->respond([
-            'data' => $this->metadataTransformer->transform($metadata)
+            'data' => $this->metadataTransformer->transform(Metadata::find($id))
         ]);
     }
 
@@ -98,8 +99,9 @@ class MetadataController extends ApiController {
     public function update($id)
     {
         try {
-            //$this->authorize('update', Metadata::find($id));
+            $this->authorize('update', Metadata::find($id));
             $this->metadataCreator->update(Input::all(), $id);
+
             return $this->respondCreated('Metadata successfully updated!');
         }
         catch(ValidationException $e) {
@@ -110,14 +112,15 @@ class MetadataController extends ApiController {
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return Response
      */
     public function destroy($id)
     {
         try {
-            //$this->authorize('destroy', Metadata::find($id));
+            $this->authorize('destroy', Metadata::find($id));
             $this->metadataCreator->destroy($id);
+
             return $this->respondCreated('Metadata successfully destroyed!');
         }
         catch(ValidationException $e) {
